Put monkeypatchs in the SpyderPdb file. Change place of docstring
